% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClassBinEBARS}
\alias{ClassBinEBARS}
\title{bivariate extended Bayesian adaptive regression spline}
\description{
a class of \code{\link[=binebars]{binebars()}}
}
\section{Fields}{

\describe{
\item{\code{new}}{constructor, see \code{binebars}}

\item{\code{rjmcmc}}{reversible jump MCMC, a wrapper is \code{mcmc}}

\item{\code{predict}}{predict by surface spline regression with EBARS}

\item{\code{knots}}{return estimated knots}
}}

