% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClassEBARS}
\alias{ClassEBARS}
\title{extended Bayesian adaptive regression spline}
\description{
a class of \code{\link[=ebars]{ebars()}}
}
\section{Fields}{

\describe{
\item{\code{new}}{constructor, see \code{ebars}}

\item{\code{rjmcmc}}{reversible jump MCMC, a wrapper is \code{mcmc}}

\item{\code{predict}}{predict by spline regression with EBARS}

\item{\code{knots}}{return estimated knots}

\item{\code{samples}}{return posterior samples}
}}

