% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClassMEBARS}
\alias{ClassMEBARS}
\title{extended Bayesian adaptive regression multivariate spline}
\description{
a cpp class
}
\section{Fields}{

\describe{
\item{\code{new}}{constructor, see \code{mebars}}

\item{\code{rjmcmc}}{run reversible jump MCMC algorithm}

\item{\code{predict}}{predict posterior response values for new data}

\item{\code{knots}}{get posterior samples of knots}

\item{\code{coefs}}{get posterior samples of regression coefficients}

\item{\code{resids}}{get posterior samples of residual standard deviations}
}}

