% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClassTriEBARS}
\alias{ClassTriEBARS}
\title{trivariate extended Bayesian adaptive regression spline}
\description{
a class of \code{\link[=triebars]{triebars()}}
}
\section{Fields}{

\describe{
\item{\code{new}}{constructor, see \code{triebars}}

\item{\code{rjmcmc}}{reversible jump MCMC, a wrapper is \code{mcmc}}

\item{\code{predict}}{predict by trivariate spline regression with EBARS}

\item{\code{knots}}{return estimated knots}

\item{\code{samples}}{return posterior samples}
}}

