% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ebars}
\alias{ebars}
\title{extended Bayesian adaptive regression univariate spline}
\usage{
ebars(
  x,
  y,
  xmin = NULL,
  xmax = NULL,
  gamma = 1,
  c = 0.3,
  times = 3,
  k = -1,
  n = -1,
  degree = 3,
  intercept = TRUE,
  npart = NULL,
  eps = 1e-08
)
}
\arguments{
\item{x}{a numeric vector, indicates the predictor values.}

\item{y}{a numeric vector, indicates the response values, the same length as x.}

\item{xmin}{double, the minimum value of predictor, default is \code{min(x)}.}

\item{xmax}{double, the maximum value of predictor, default is \code{max(x)}.}

\item{gamma}{double, the exponent in extended BIC, default value is \code{1.0}.
It should be constrained to the range \verb{[0.5, 2.0]}.}

\item{c}{double, the constant affects the proposal distribution in MCMC,
default value is \code{0.3}. It has to be contained in \verb{[0,1]}.}

\item{times}{double, decides the number of all potential knots if \code{n<0}.
In this case, the number is \code{length(x) * times}, default value is \code{3.0}.}

\item{k}{int, the number of knots is fixed if \code{k>0}. Otherwise \code{k} will
be estimated in the algorithm, default value is \code{-1}.}

\item{n}{int, the number of all potential knots, default value is \code{-1}.}

\item{degree}{int, the degree of polynomial spline, default value is \code{3}.}

\item{intercept}{bool, whether the intercept is included in the basis,
default value is \code{TRUE}.}

\item{npart}{int, minimum number of non-zero points in a basis function.
Defaults to \code{20} or \code{0.1 * length(x)}, whichever is smaller.}

\item{eps}{double, a small constant to define "non-zero", default value is \code{1e-8}.}
}
\value{
An R6 object of class EBARS with the following methods:
\describe{
\item{\code{rjmcmc(burns, steps)}}{Run reversible jump MCMC algorithm.
\itemize{
\item \code{burns}: Number of burn-in iterations, standard value is \code{5000}.
\item \code{steps}: Number of posterior sampling iterations, standard value is \code{5000}.
}}
\item{\code{predict(x_new)}}{Predict response values for new data.
\itemize{
\item \code{x_new}: Numeric vector of new predictor values
\item Returns: Matrix of predictions, each column is a posterior sample
}}
\item{\code{knots()}}{Get posterior samples of knot locations.
\itemize{
\item Returns: List of knot vectors from each MCMC iteration
}}
\item{\code{coefs()}}{Get posterior samples of regression coefficients.
\itemize{
\item Returns: List of coefficient vectors from each MCMC iteration
}}
\item{\code{resids()}}{Get posterior samples of residual standard deviations.
\itemize{
\item Returns: List of sigma values from each MCMC iteration
}}
}
}
\description{
see \link{ClassEBARS} for more details about class EBARS
}
\examples{
library(ABMRS)
library(splines)
set.seed(1234)
knot = c(0.4,0.4,0.4,0.4,0.7)
beta = matrix(c(2,-5,5,2,-3,-1,2),ncol=1)
m = 200

# generate train data set
x = c(0,sort(runif(m, 0, 1))[c(-1,-m)],1)
B = ns(x,knots=knot,intercept=TRUE,Boundary.knots=c(0,1))
y = B\%*\%beta
y_h = y + rnorm(m,0,0.05)

# run EBARS
a = ebars(x,y_h,times = 2)
a$rjmcmc(burns=1000,steps=1000)

# generate test data set
m_new = 50
x_new = runif(m_new,0,1)
y_new = predict(B,x_new)\%*\%beta
pred = a$predict(x_new)
y_hat = rowMeans(pred)

}
