% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_fitting.R
\name{manifold_fitting}
\alias{manifold_fitting}
\title{Fit an 1 or 2 dimensional manifold with EBARS}
\usage{
manifold_fitting(x, d, epsilon = 0.05, max_iter = 100, k = 10)
}
\arguments{
\item{x}{a numeric matrix, (n,p), each row indicates a point in R^p.}

\item{d}{int, the intrinsic dimension.}

\item{epsilon}{double, the stop criterion, default to \code{0.05}.}

\item{max_iter}{int, the maximum iteration number, default to \code{100}.}

\item{k}{int, number of shortest dissimilarities retained for a point in
\code{vegan::isomap()}, default to \code{10}.}
}
\value{
a list including:
\itemize{
\item{\code{map}, the estimated embedding map.}
\item{\code{lower}, the lower bound of covariates in \code{map}.}
\item{\code{upper}, the upper bound of covariates in \code{map}.}
}
}
\description{
Fit an 1 or 2 dimensional manifold with EBARS
}
\examples{
manifold=function(t){ return(c(cos(t),sin(t)))}
I=100
t=runif(I,min = 0,max = 1.5*pi)
X=manifold(t)
sd.noise=0.05
e1=rnorm(I,mean = 0, sd=sd.noise)
e2=rnorm(I,mean = 0, sd=sd.noise)
data.points=X+cbind(e1,e2)
res = manifold_fitting(x=data.points,d=1)
center = (res$lower+res$upper)/2
radius = abs(res$upper-res$lower)/2
x_test = res$map(seq(center-radius*0.95,center+radius*0.95,length.out=500))
plot(data.points[,1], data.points[,2],
pch=20, col="grey",
main="Principal Manifold Estimation",
xlab=" ", ylab=" ")
lines(x_test[,1],x_test[,2],col="red",type = "l",lwd=3)
}
