% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mebars}
\alias{mebars}
\title{extended Bayesian adaptive regression multivariate spline}
\usage{
mebars(
  x,
  y,
  gamma = 1,
  c = 0.3,
  times = NULL,
  ks = NULL,
  ns = NULL,
  degrees = NULL,
  intercepts = NULL
)
}
\arguments{
\item{x}{a numeric matrix, indicates the predictor values, each column is a predictor.}

\item{y}{a numeric vector, indicates the response values.}

\item{gamma}{double, the exponent in extended BIC, default value is \code{1.0}.
It should be greater than 0.}

\item{c}{double, the constant affects the proposal distribution in MCMC,
default value is \code{0.3}. It has to be contained in \verb{[0,1]}.}

\item{times}{a double vector with the length \code{ncol(x)}, decides the number of all potential knots if \code{ns<0}.
In this case, the number is \code{nrow(x) * times}, default value is \code{c(3, 3, ...)}.}

\item{ks}{an integer vector with the length \code{ncol(x)}, the number of knots is fixed if \code{ks>0}. Otherwise \code{ks} will
be estimated in the algorithm, default value is \code{c(-1, -1, ...)}.}

\item{ns}{an integer vector with the length \code{ncol(x)}, the number of all potential knots, default value is \code{c(-1, -1, ...)}.}

\item{degrees}{an integer vector with the length \code{ncol(x)}, the degree of polynomial spline, default value is \code{c(3, 3, ...)}.}

\item{intercepts}{a logical vector with the length \code{ncol(x)}, whether the intercept is included in the basis,
default value is \code{c(TRUE, TRUE, ...)}.}
}
\value{
An S4 object of class MEBARS with the following methods:
\describe{
\item{\code{rjmcmc(burns, steps)}}{Run reversible jump MCMC algorithm.
\itemize{
\item \code{burns}: Number of burn-in iterations
\item \code{steps}: Number of posterior sampling iterations
}}
\item{\code{predict(x_new)}}{Predict response values for new data.
\itemize{
\item \code{x_new}: Numeric matrix of new predictor values
\item Returns: Matrix of predictions, each column is a posterior sample
}}
\item{\code{knots()}}{Get posterior samples of knot locations.
\itemize{
\item Returns: List of knot vectors from each MCMC iteration
}}
\item{\code{coefs()}}{Get posterior samples of regression coefficients.
\itemize{
\item Returns: List of coefficient vectors from each MCMC iteration
}}
\item{\code{resids()}}{Get posterior samples of residual standard deviations.
\itemize{
\item Returns: List of sigma values from each MCMC iteration
}}
}
}
\description{
see \link{ClassMEBARS} for more details about class MEBARS
}
\examples{
library(ABMRS)
library(splines)
set.seed(1234)
# tensor spline
beta = matrix(rnorm(40,0,1),ncol=1)
fss <- function(x,y){
  xix = c(0.2,0.3)
  xiy = c(0.5,0.5,0.5,0.5,0.7)
  B = tensor_spline(cbind(x,y),list(xix,xiy), c(3,3), c(FALSE,FALSE))
  return(B\%*\%beta)
}
# parameters' configuration
m_train = 1000; m_test = 200
noise = 0.1
# generate train data set
x_1 = c(runif(m_train-2,0,1),0,1)
x_2 = c(runif(m_train-2,0,1),0,1)
y = fss(x_1,x_2)
y_h = y + rnorm(m_train,0,noise)
# generate test set
x_1_new = runif(m_test,0,1)
x_2_new = runif(m_test,0,1)
y_new = fss(x_1_new,x_2_new)
# run suface EBARS
time_start = Sys.time()
my_mebars = mebars(cbind(x_1,x_2), y_h)
my_mebars$rjmcmc(100,100)
time_end = Sys.time()
time_end - time_start
pred = my_mebars$predict(cbind(x_1_new,x_2_new))
y_hat = rowMeans(pred)
mean((y_new-y_hat)^2)

}
