% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spline}
\alias{spline}
\title{create a univariate B-spline basis matrix}
\usage{
spline(x, xi, degree, intercept)
}
\arguments{
\item{x}{a numeric vector of predictor values.}

\item{xi}{a numeric vector, indicates the knots.}

\item{degree}{int, the degree of polynomial, default value is \code{3}.}

\item{intercept}{bool, whether the intercept is included in the basis, default value is \code{FALSE}.}
}
\value{
a B-spline basis matrix.
}
\description{
create a univariate B-spline basis matrix
}
