% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tri_tensor_spline}
\alias{tri_tensor_spline}
\title{create a trivariate tensor product spline basis matrix}
\usage{
tri_tensor_spline(
  x,
  xi_1,
  xi_2,
  xi_3,
  degree_1 = 3L,
  degree_2 = 3L,
  degree_3 = 3L,
  intercept_1 = FALSE,
  intercept_2 = FALSE,
  intercept_3 = FALSE
)
}
\arguments{
\item{x}{a numeric matrix, (m,3), each row indicates a predictor value (x_1,x_2,x_3).}

\item{xi_1}{a numeric vector, indicates \code{k_1} knots in x_1.}

\item{xi_2}{a numeric vector, indicates \code{k_2} knots in x_2}

\item{xi_3}{a numeric vector, indicates \code{k_3} knots in x_3}

\item{degree_1}{int, the degree of polynomial in x_1, default value is \code{3}.}

\item{degree_2}{int, the degree of polynomial in x_2, default value is \code{3}.}

\item{degree_3}{int, the degree of polynomial in x_3, default value is \code{3}.}

\item{intercept_1}{bool, whether the intercept is included in the basis in x_1,
default value is \code{FALSE}.}

\item{intercept_2}{bool, whether the intercept is included in the basis in x_2,
default value is \code{FALSE}.}

\item{intercept_3}{bool, whether the intercept is included in the basis in x_3,
default value is \code{FALSE}.}
}
\value{
a trivariate tensor product spline basis matrix.
}
\description{
create a trivariate tensor product spline basis matrix
}
