% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ebars}
\alias{ebars}
\title{create an instance of EBARS}
\usage{
ebars(x, y, gamma = 1, c = 0.3, times = 2, n = -1)
}
\arguments{
\item{x}{a numeric vector, indicates the predictor values.}

\item{y}{a numeric vector, indicates the response values, the same length as x.}

\item{gamma}{double, the exponent in extended BIC, default value is \code{1.0}.
It should be contained in \verb{[0,1]} (Nevertheless, outside is valid).}

\item{c}{double, the constant affects the proposal distribution in MCMC,
default value is \code{0.3}. It has to be contained in \verb{[0,1]}.}

\item{times}{int, decides the number of all potential knots if \code{n<0}.
In this case, the number is \code{length(x) * times}, default value is \code{2}.}

\item{n}{int, the number of all potential knots, default value is \code{-1}.}
}
\value{
An exposed R class of cpp class called EBARS.
}
\description{
create an instance of EBARS
}
\examples{
library(EBARS)
set.seed(1234)
knot = c(0.4,0.4,0.4,0.4,0.7)
beta = matrix(c(2,-5,5,2,-3,-1,2),ncol=1)
m = 200

# generate train data set
x = c(0,sort(runif(m_train, 0, 1))[c(-1,-m)],1)
B = ns(x,knots=knot,intercept=TRUE,Boundary.knots=c(0,1))
y = B\%*\%beta
y_h = y + rnorm(m_train,0,0.05)

# run EBARS
a = ebars(x,y_h,c=0.3,times = 2)
a$mcmc()
print(a$knots())

# generate test data set
m_new = 50
x_new = runif(m_new,0,1)
y_new = predict(B,x_new)\%*\%beta
y_pred = a$predict(x_new)

}
