% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tensor_spline}
\alias{tensor_spline}
\title{create a bivariate tensor product spline basis matrix}
\usage{
tensor_spline(x, xi_1, xi_2)
}
\arguments{
\item{x}{a numeric matrix, (m,2), each row indicates a predictor value (x_1,x_2).}

\item{xi_1}{a numeric vector, indicates \code{k_1} knots in x_1.}

\item{xi_2}{a numeric vector, indicates \code{k_2} knots in x_2}
}
\value{
a bivariate tensor product spline basis matrix, m rows and (k_1+3)*(k_2+3) cols.
}
\description{
create a bivariate tensor product spline basis matrix
}
