% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tensor_spline}
\alias{tensor_spline}
\title{create a general tensor product spline basis matrix for arbitrary dimensions}
\usage{
tensor_spline(x, xis, degrees, intercepts, xmin, xmax)
}
\arguments{
\item{x}{a numeric matrix, (m,d), each row indicates a predictor value.}

\item{xis}{a list of numeric vectors, each element contains knots for one dimension.}

\item{degrees}{an integer vector, degrees for each dimension.}

\item{intercepts}{a logical vector, whether intercepts are included for each dimension.}

\item{xmin}{a numeric vector, minimum boundary for each input dimension.}

\item{xmax}{a numeric vector, maximum boundary for each input dimension.}
}
\value{
a tensor product B-spline basis matrix.
}
\description{
create a general tensor product spline basis matrix for arbitrary dimensions
}
\examples{
# 2D example
x <- matrix(runif(100), ncol=2)
xmin <- c(0,0)
xmax <- c(1,1)
xis <- list(c(0.3, 0.6), c(0.4, 0.5))
B <- tensor_spline(x, xis, c(3,3), c(FALSE,FALSE), xmin, xmax)

# 3D example
x <- matrix(runif(150), ncol=3)
xmin <- c(0,0,0)
xmax <- c(1,1,1)
xis <- list(c(0.2, 0.5), c(0.3, 0.7), c(0.4))
B <- tensor_spline(x, xis, c(3,3,3), c(TRUE,TRUE,TRUE), xmin, xmax)
}
